Changed plotting routines to use figure and plot_grid
